+++ checkpoint/orbax/checkpoint/tree/__init__.py
@@ -29,4 +29,5 @@ from orbax.checkpoint.tree.utils import (
     to_flat_dict,
     to_shape_dtype_struct,
     tuple_path_from_keypath,
+    get_param_names,
 )